Added CL_PROGRAM_BUILD_LOG. Fixed bug in Example1.cl (some compilers): pown(src[id],4)->(int)pown((float)src[id],4)

+++ Example1/Example1.cl
@@ -1,5 +1,5 @@
 __kernel void somekernel(__global int *src,__global int *dst) 
 {
 	int id = get_global_id(0);
-	dst[id] = pown(src[id],4);
+	dst[id] = (int)pown((float)src[id],4);
 }
